Support for Big SSD
-Tested with up to 2TB disk size
-Added support for Delay mode:
    1.real - backwards compatible, same delay that was
      implemented until now
    2.emulated - no delay, execution of test continues normally,
      delay is only logged.
-All relevant variable in code were updated to the appropriate kind,
 size of variable was increased to accommodate larger disk, and
 signed/unsigned was changed depending on actual use of variable
-Calculation of ssd_stats: utilization and read/write speed was fixed
 to account for multiple rt_analyzers
-Added ssd_stats: occupied_pages, logical_write_count,
 block_erase_count, channel_switch_to_write, channel_switch_to_read
-Added mechanism for syncing to log without arbitrary time sleep
-Updated GC: Victim list is now sorted by valid_page_nb, accelerating
 GC algorithm for large disk sizes.
-Update legacy tests for new changes
-ssd_write_read test now has tests for checking all ssd_stats
 during test of read, write, mixed read/write and GC tests

+++ eVSSIM/FTL_SOURCE/PAGE_MAP/ftl_sect_strategy.h
@@ -8,7 +8,7 @@ ftl_ret_val _FTL_WRITE_SECT(uint64_t sector_nb, unsigned int length);
 
 ftl_ret_val _FTL_READ(uint64_t sector_nb, unsigned int offset, unsigned int length);
 ftl_ret_val _FTL_WRITE(uint64_t sector_nb, unsigned int length);
-ftl_ret_val _FTL_COPYBACK(int32_t source, int32_t destination);
+ftl_ret_val _FTL_COPYBACK(uint64_t source, uint64_t destination);
 ftl_ret_val _FTL_CREATE(void);
 ftl_ret_val _FTL_DELETE(void);
